--- backend/bang-ggood/src/test/java/com/bang_ggood/checklist/domain/OptionTest.java
@@ -1,22 +1,0 @@
-package com.bang_ggood.checklist.domain;
-
-import com.bang_ggood.option.domain.Option;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-
-import static org.assertj.core.api.Assertions.assertThat;
-
-class OptionTest {
-
-    @DisplayName("옵션 포함 성공: 포함하는 경우")
-    @Test
-    void contains_true() {
-        assertThat(Option.contains(1)).isTrue();
-    }
-
-    @DisplayName("옵션 포함 성공 : 포함하지 않는 경우")
-    @Test
-    void contains_false() {
-        assertThat(Option.contains(9999)).isFalse();
-    }
-}
