+++ backend/bang-ggood/src/test/java/com/bang_ggood/option/domain/OptionTest.java
@@ -1,6 +1,5 @@
-package com.bang_ggood.checklist.option;
+package com.bang_ggood.option.domain;
 
-import com.bang_ggood.option.domain.Option;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
