+++ backend/bang-ggood/src/test/java/com/bang_ggood/question/domain/QuestionTest.java
@@ -1,8 +1,7 @@
-package com.bang_ggood.checklist.domain;
+package com.bang_ggood.question.domain;
 
 import com.bang_ggood.global.exception.BangggoodException;
 import com.bang_ggood.global.exception.ExceptionCode;
-import com.bang_ggood.question.domain.Question;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import java.util.Arrays;
